LDAP auth basically working, with a slight bug...
Authentication against LDAP now works, and will auto-create a user object
with the correct details if it doesn't already exist.

If it fails, it's possible to fall back to form auth or similar.

Currently there is a bug whereby login works fine but tells the user it failed. This is inconvenient.